Updated screenshot command

+++ src/commands/Fun/screenshot.js
@@ -3,20 +3,27 @@ const window = new Snowshot();
 const { MessageAttachment } = require('discord.js');
 
 module.exports.run = async (bot, message, args, settings) => {
-	try {
-		window.screenshot(args[0]).then(data => {
-			if (!data) {
-				return message.error(settings.Language, 'ERROR_MESSAGE').then(m => m.delete({ timeout: 5000 }));
-			} else {
-				const attachment = new MessageAttachment(data, 'website.png');
-				message.channel.send(attachment);
-			}
-		});
-	} catch (err) {
+	// make sure a website was entered
+	if (!args[0]) {
 		if (message.deletable) message.delete();
-		bot.logger.error(err.message);
-		return message.error(settings.Language, 'ERROR_MESSAGE').then(m => m.delete({ timeout: 5000 }));
+		return message.error(settings.Language, 'INCORRECT_FORMAT', bot.commands.get('screenshot').help.usage.replace('${PREFIX}', settings.prefix)).then(m => m.delete({ timeout: 5000 }));
 	}
+
+	// send 'waiting' message
+	const msg = await message.channel.send('Creating screenshot of website.');
+
+	// try and create screenshot
+	window.screenshot(args[0]).then(async data => {
+		if (!data) {
+			if (message.deletable) message.delete();
+			msg.delete();
+			return message.error(settings.Language, 'ERROR_MESSAGE').then(m => m.delete({ timeout: 5000 }));
+		} else {
+			const attachment = new MessageAttachment(data, 'website.png');
+			await message.channel.send(attachment);
+			msg.delete();
+		}
+	});
 };
 
 module.exports.config = {
